add user ia and improve displays
* added an user ia which read an character in input and make the player play
the column according to the input if this one is valid.
keep asking for a good column if not.
this ia can display text instructions or display the game depending on the
tracing parameters activated

* added two test using this ia to play alone vs iatest1 (always playing
column 1) or vs another user. They weren't added to all the test because
they cannot be test by the computer only.

* improving the display in gameLauncher by displaying the whole game after
each move instead of just the last move

* put new tracing constants in testUtil and disabled all of them