Disables some gameOver UnitTests if numberOfAlignedPawnsToWin/1 is changed. #3797
Some unit tests of gameOver module rely on the fact that we need 4
aligned pawns to win. This is because test grids are generated manually,
so we cannot adapt to other settings. In order to avoid the tests to
fail in case of default value changed (other than 4), the specific
subset of unit tests relying on this number are automatically disabled.
A notice message will be printed to inform the developer that some unit
tests have been skipped.